fixed bugs where objects would not be affected according to the API c…

+++ src/main/java/docSharing/service/AuthService.java
@@ -60,7 +60,7 @@ public class AuthService {
      * Method used after a user clicks on the link he got on email.
      * @param id - user email
      */
-    public int activate(Integer id) {
+    public int activate(Long id) {
         // check if id exists
         return userRepository.updateIsActivated(true, id);
     }
